NOISSUE eliminate ProgressProvider

+++ logic/java/JavaVersionList.cpp
@@ -178,7 +178,7 @@ void JavaListLoadTask::executeTask()
 
 	m_job = std::shared_ptr<JavaCheckerJob>(new JavaCheckerJob("Java detection"));
 	connect(m_job.get(), SIGNAL(finished(QList<JavaCheckResult>)), this, SLOT(javaCheckerFinished(QList<JavaCheckResult>)));
-	connect(m_job.get(), SIGNAL(progress(int, int)), this, SLOT(checkerProgress(int, int)));
+	connect(m_job.get(), SIGNAL(progress(qint64,qint64)), this, SLOT(checkerProgress(qint64, qint64)));
 
 	qDebug() << "Probing the following Java paths: ";
 	int id = 0;
@@ -197,7 +197,7 @@ void JavaListLoadTask::executeTask()
 	m_job->start();
 }
 
-void JavaListLoadTask::checkerProgress(int current, int total)
+void JavaListLoadTask::checkerProgress(qint64 current, qint64 total)
 {
 	float progress = (current * 100.0) / total;
 	this->setProgress((int) progress);
